+++ src/actions_impl.h
@@ -25,18 +25,16 @@
 #include "firebolt/actions.h"
 #include <firebolt/helpers.h>
 
-namespace Firebolt::Actions
-{
+namespace Firebolt::Actions {
 
-class ActionsImpl : public IActions
-{
+class ActionsImpl : public IActions {
 public:
     explicit ActionsImpl(Firebolt::Helpers::IHelper& helper);
     ActionsImpl(const ActionsImpl&) = delete;
     ActionsImpl& operator=(const ActionsImpl&) = delete;
     ~ActionsImpl() override = default;
 
-    Result<void> intent(const std::string& intent) const override;
+    Result<std::string> intent() const override;
 
     Result<SubscriptionId> subscribeOnIntent(std::function<void(const std::string&)>&& notification) override;